Move event tie counting into fossil collection
The previous method for counting event ties on each PE was done
during scheduling by checking to see if the current event being
processed has the same timestamp as the minimum in the queue.

This will result in overcounting as it's possible for the 'same
event' to be in the queue multiple times but with different
event IDs. Also this would mean that an event with the same
timestamp but a different destination LP would qualify as an
event tie. This will have zero effect on determinism or model
validity as those two events would be uncorrelated with each
other despite qualifying as ties.

If we wanted to know how many significant event ties existed
in the simulation, this check should be done during fossil
collection. That way the same event existing in the queue
more than once doesn't throw off the count.